Working OpenAI backend
Instead of raw text, for compatability with the chat api, we convert the nodes to messages.

A mixed-quality `openai_rust` crate was used for the backend. Really it's not horrible and it supports the streaming API but it require tokio and as a result it stands in the way of a web port.

We may wish to find another crate or write one using reqwests.