add tree nav bar
This adds a tree navigation bar like in weave. Like elsewhere in `weave` the same underlying data is shared between multiple ui elements.

A design philosphy we follow with this app is to have a single source of truth and never repeat ourselves. This kills lots and lots of bugs dead.

While writing this, I realized the auto-collapse could apply not just to the tree view, but to the nodes as well. It might be better to move that setting to the App and make it persistent so you don't have to click the button every time you want to collapse. The only issue is selecting nodes when collapsed is not possible (because the select button would be collpsed), so the path could not be changed. This isn't really desirable and it would require changing how selection works. In response to `clicked()` could probably work, and it would make the select button unnecessary. This is probably more natural anyway.