Add initial "change topology" support
Nodes can now be copied and pasted with an internal clipboard. What doesn't work yet is the author metadata.

Nodes *can* be copied and pasted betwee stories but the author's id will be reset to the head's.

This is the downside to storing the authors in the story and not the nodes themselves. It might be worth it to have a database of authors in the app itself. This is one way to solve it without copying the data on every single node.