Force directed layout WIP
It mostly works, but there are some issues that will be resolved by debugging and visualzation.

The general idea was taken from reading about how Gephi does it. Nodes repel each other with inverse gravity and attact each other with spring forces, linearly.

To reduce time complexity, forces are only calculated between parent and child as well as immediate children. There's also a child centroid the parent is attracted to.

All this *mostly* works but there are bugs so it's not ready yet for merge into `dev`.

Outstanding issues are:
* nodes flying off.
* colissions not quite working right
* nodes don't move when they should.

Solving these is a matter of visualizing what's going on. Node velocity is already visualized with a red outline in debug builds. The centroids should also be visualized.

Also a bunch of typos have been corrected.