Improved gravity
The gravitational centroid is now a weighted average of the local and global. This seems to work better. Also fixed is that we were applying gravity twice.

Also solved the issue of the nodes not moving. It was, as suspected, that egui rounds the positions of the windows, so now we only set the position of the node to the window position if the node is being dragged.

Otherwise force has to get really high to get the nodes moving, so in most cases they don't.

Because damping wasn't the issue, damping has been set higher. Using it as a cutoff value also works as intended and nodes that are below a certain velocity stop and are no longer redrawn.

Issues still are:
* the root is never comes to a full stop, so the `redraw` optimization doesn't work perfectly. Even when the simulation fully converges every frame is still drawn which is not what we want.
* Nodes still somtime fly off, sometimes even outside of the window bounds which shouldn't be possible for multiple reasons but happens. Need to review the value clampings probably.