Fix potential crash
I don't know how, exactly, but the active Story can be removed without setting `App::active_story` to `None` leading to a crash.

If there is a bug, it's here:
https://github.com/mdegans/weave/blob/4019e18d7413206b08c3a39edd482919d181e63b/src/app.rs#L721

But there active_story is set to None in the case where the active story is deleted.

And data races shouldn't be an issue because of the immediate mode nature of the gui. In any case, this ensures that if it *is* an invalid index it won't panic.