Fix force directed layout speed
Fix Force directed layout slower on 60hz displays #16

The time step of the force-directed layout is now based on the current frame rate.

Dragging a window between a 60hz display and a 120hz display now works propertly. This should even account for frame dropping.

The time step is smoothed with each new step being averaged with the last acounting for frame pacing issues and jitter.

Additionally, a speed slider has been added to the UI to control a multiplier to the time step.