Configure `drama_llama::Engine` to use all threads
By default `drama_llama` uses only one thread for CPU work (IIRC). This could be changed but for the moment we'll set it here to use all threads. Perhaps we'll add a widget to make it configurable.

This will addresss the slowness mentioned in "crash on unsupported model #6"

It's *almost* as fast as GPU-only when using all cores on Metal.

This mostly fixes the performance hit for models quantized with `--leave-output-tensor` in which some operands are bf16 or other unsupported types.

Additionally this fixes an issue where two engine existed at once by dropping one before creating another.